working on numreplies

+++ Numerics.cpp
@@ -6,7 +6,7 @@ using namespace irc;
 #define END "\r\n"
 
 void	sendNumeric(int fd, std::string msg) {
-	int bytes_send = send(fd, msg.c_str(), msg.size(), 0);
+	unsigned long bytes_send = send(fd, msg.c_str(), msg.size(), 0);
 	if (bytes_send < msg.size()) {
 		std::cerr << "Error while sending message to fd = " << fd << std::endl;
 		// TODO close connection
